+++ resources/views/components/form/input-conditional-checkbox.blade.php
@@ -7,8 +7,9 @@
     'model' => '',
 ])
 
-<div x-data="{ show{{ $id }}: @entangle($attributes->wire('model')) }">
-    <div class="pl-3 pb-2 mt-2 space-y-4">
+<div class="pl-3 pb-2 mt-2 space-y-4"  x-data="{ show{{ $id }}: @entangle($attributes->wire('model')) }">
+    <!-- div class="pl-3 pb-2 mt-2 space-y-4"-->
+    <div>
         <x-form.input-checkbox
             model="{{ $model }}"
             name="{{ $name }}"
@@ -18,7 +19,8 @@
             >
         </x-form.input-checkbox>
     </div>
-    <div x-show="show{{ $id }}">
+
+    <div class="pl-5 pt-2 pb-2 pr-5 bg-blue-50 rounded-2xl" x-show="show{{ $id }}">
         {{ $slot }}
 
     </div>